+++ app/api/registrar_evento/route.ts
@@ -1,16 +1,6 @@
-import { Endereco } from "@/app/lib/Endereco";
 import { Evento } from "@/app/lib/Evento";
 import { Status } from "@/generated/prisma/enums";
 import { parseJsonBody } from "@/app/api/lib/request";
-
-type EnderecoPayload = {
-  cidade: string;
-  bairro: string;
-  rua: string;
-  cep: string;
-  apartamento?: string;
-  numero?: string;
-};
 
 type RegistrarEventoPayload = {
   nome: string;
@@ -18,7 +8,7 @@ type RegistrarEventoPayload = {
   data: string;
   horarioInicio: string;
   horarioFim: string;
-  endereco: EnderecoPayload;
+  local: string;
   vagasDisponiveis: number;
   status: Status;
   organizador: number;
@@ -35,7 +25,6 @@ function isValidPayload(payload: unknown): payload is RegistrarEventoPayload {
   if (typeof payload !== "object" || payload === null) return false;
 
   const body = payload as Record<string, unknown>;
-  const endereco = body.endereco as Record<string, unknown> | undefined;
 
   return (
     typeof body.nome === "string" &&
@@ -48,19 +37,8 @@ function isValidPayload(payload: unknown): payload is RegistrarEventoPayload {
     body.horarioInicio.trim().length > 0 &&
     typeof body.horarioFim === "string" &&
     body.horarioFim.trim().length > 0 &&
-    typeof endereco === "object" &&
-    endereco !== null &&
-    typeof endereco.cidade === "string" &&
-    endereco.cidade.trim().length > 0 &&
-    typeof endereco.bairro === "string" &&
-    endereco.bairro.trim().length > 0 &&
-    typeof endereco.rua === "string" &&
-    endereco.rua.trim().length > 0 &&
-    typeof endereco.cep === "string" &&
-    endereco.cep.trim().length > 0 &&
-    (endereco.apartamento === undefined ||
-      typeof endereco.apartamento === "string") &&
-    (endereco.numero === undefined || typeof endereco.numero === "string") &&
+    typeof body.local === "string" &&
+    body.local.trim().length > 0 &&
     typeof body.vagasDisponiveis === "number" &&
     Number.isFinite(body.vagasDisponiveis) &&
     body.vagasDisponiveis >= 0 &&
@@ -101,25 +79,12 @@ export async function POST(request: Request) {
     );
   }
 
-  const endereco = new Endereco(
-    payload.endereco.cidade,
-    payload.endereco.bairro,
-    payload.endereco.rua,
-    payload.endereco.cep,
-    payload.endereco.apartamento,
-    payload.endereco.numero
-  );
-  const createdEndereco = await endereco.storeOnDb();
-
   const evento = new Evento(
     payload.organizador,
     statusToPublicado(payload.status),
     payload.status,
-    payload.nome,
-    payload.horarioInicio,
-    payload.horarioFim,
     dataHoraInicio,
-    createdEndereco.idendere_o,
+    undefined,
     payload.descricao,
     payload.vagasDisponiveis
   );
