+++ app/page.tsx
@@ -7,7 +7,6 @@ export default function Home() {
     <div className="flex flex-col flex-1 items-center justify-center bg-zinc-50 font-sans dark:bg-black">
       <Link href="/pages/organizador/registrar_evento">Criar novo evento</Link>
       <Link href="/perfil">Perfil</Link>
-      <Link href="/eventos-disponiveis">Eventos</Link>
       <main className="flex flex-1 w-full max-w-3xl flex-col items-center justify-between py-32 px-16 bg-white dark:bg-black sm:items-start">
         <Image
           className="dark:invert"
